+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Inputs size to small by default [#556](https://github.com/CartoDB/carto-react/pull/556)
 - New color variant in FAB button [#554](https://github.com/CartoDB/carto-react/pull/554)
 - Adjustments in theme to fix uses cases in cloud-native [#553](https://github.com/CartoDB/carto-react/pull/553/)
 - Default props of core components to meet the common use case [#552](https://github.com/CartoDB/carto-react/pull/552/)